Generative_AII/requirements.txt
Flask
torch
transformers
diffusers
Pillow
